change client pos
